tidy up the project
