+++ lib/jekyll-admin/public/src/actions/datafiles.js
@@ -11,7 +11,6 @@ import {
 import { validationError } from '../actions/utils';
 
 import { get, put, del } from '../utils/fetch';
-import { toJSON } from '../utils/helpers';
 import { validator } from '../utils/validation';
 
 export function fetchDataFiles() {
@@ -54,13 +53,10 @@ export function putDataFile(filename, data) {
     if (errors.length) {
       return dispatch(validationError(errors));
     }
-    const body = {
-      raw_content: data
-    };
     dispatch({type: ActionTypes.CLEAR_ERRORS});
     return put(
       putDataFileUrl(filename),
-      JSON.stringify(body),
+      JSON.stringify({ raw_content: data }),
       { type: ActionTypes.PUT_DATAFILE_SUCCESS, name: "file"},
       { type: ActionTypes.PUT_DATAFILE_FAILURE, name: "error"},
       dispatch
