+++ lib/jekyll-admin/public/src/actions/tests/collections.spec.js
@@ -11,13 +11,14 @@ import { doc, new_doc, documents, collections } from './fixtures';
 
 const middlewares = [ thunk ];
 const mockStore = configureMockStore(middlewares);
+const filename = doc.path.substring(doc.path.lastIndexOf('/') + 1);
 
 describe('Actions::Collections', () => {
   afterEach(() => {
     nock.cleanAll();
   });
 
-  it('fetches collections', () => {
+  it('fetches collections successfully', () => {
     nock(API)
       .get('/collections')
       .reply(200, collections);
@@ -35,7 +36,7 @@ describe('Actions::Collections', () => {
       });
   });
 
-  it('fetches the collection', () => {
+  it('fetches the collection successfully', () => {
     let collection_name = "movies";
     nock(API)
       .get(`/collections/${collection_name}`)
@@ -54,8 +55,8 @@ describe('Actions::Collections', () => {
       });
   });
 
-  it('fetches the collection documents', () => {
-    let collection_name = "posts";
+  it('fetches the collection documents successfully', () => {
+    const collection_name = "posts";
     nock(API)
       .get(`/collections/${collection_name}/documents`)
       .reply(200, documents);
@@ -73,11 +74,9 @@ describe('Actions::Collections', () => {
       });
   });
 
-  it('fetches the document', () => {
-    let collection_name = "movies";
-    let id = "the-revenant";
+  it('fetches the document successfully', () => {
     nock(API)
-      .get(`/collections/${collection_name}/${id}`)
+      .get(`/collections/${doc.collection}/${filename}`)
       .reply(200, doc);
 
     const expectedActions = [
@@ -87,14 +86,13 @@ describe('Actions::Collections', () => {
 
     const store = mockStore({ currentDocument: {} });
 
-    return store.dispatch(actions.fetchDocument(collection_name, id))
+    return store.dispatch(actions.fetchDocument(doc.collection, filename))
       .then(() => {
         expect(store.getActions()).toEqual(expectedActions);
       });
   });
 
-  it('deletes the document', () => {
-    const filename = doc.path.substring(doc.path.lastIndexOf('/') + 1);
+  it('deletes the document successfully', () => {
     nock(API)
       .delete(`/collections/${doc.collection}/${filename}`)
       .reply(200, { message: 'Deleted' });
@@ -112,7 +110,6 @@ describe('Actions::Collections', () => {
   });
 
   it('creates DELETE_DOCUMENT_FAILURE when deleting document failed', () => {
-    const filename = doc.path.substring(doc.path.lastIndexOf('/') + 1);
     nock(API)
       .delete(`/collections/${doc.collection}/${filename}`)
       .replyWithError('something awful happened');
@@ -130,7 +127,7 @@ describe('Actions::Collections', () => {
       });
   });
 
-  it('updates the document', () => {
+  it('updates the document successfully', () => {
     nock(API)
       .put(`/collections${doc.id+doc.ext}`)
       .reply(200, doc);
@@ -168,7 +165,7 @@ describe('Actions::Collections', () => {
       });
   });
 
-  it('creates the document', () => {
+  it('creates the document successfully', () => {
     nock(API)
       .put(`/collections/${new_doc.collection}/${new_doc.path}`)
       .reply(200, doc);
@@ -188,10 +185,13 @@ describe('Actions::Collections', () => {
 
   it('creates VALIDATION_ERROR if required field is not provided.', () => {
     const expectedActions = [
-      { type: types.VALIDATION_ERROR, errors: [
-        'The title is required.',
-        'The filename is required.'
-      ] }
+      {
+        type: types.VALIDATION_ERROR,
+        errors: [
+          'The title is required.',
+          'The filename is required.'
+        ]
+      }
     ];
 
     const store = mockStore({metadata: { metadata: _.omit(doc, ['title','path']) }});
@@ -199,5 +199,4 @@ describe('Actions::Collections', () => {
     store.dispatch(actions.putDocument(doc.id, doc.collection));
     expect(store.getActions()).toEqual(expectedActions);
   });
-
 });
