+++ lib/jekyll-admin/public/src/actions/tests/config.spec.js
@@ -9,32 +9,32 @@ import expect from 'expect';
 const middlewares = [ thunk ];
 const mockStore = configureMockStore(middlewares);
 
+import { config } from './fixtures';
+
 describe('Actions::Config', () => {
   afterEach(() => {
     nock.cleanAll();
   });
 
-  it('creates FETCH_CONFIG_SUCCESS when fetching config has been done', () => {
+  it('fetches configuration successfully', () => {
     nock(API)
       .get('/configuration')
-      .reply(200, { title: 'Awesome Title' });
+      .reply(200, config);
 
     const expectedActions = [
       { type: types.FETCH_CONFIG_REQUEST },
-      { type: types.FETCH_CONFIG_SUCCESS, config: { title: 'Awesome Title' } }
+      { type: types.FETCH_CONFIG_SUCCESS, config }
     ];
 
     const store = mockStore({ config: {} });
 
     return store.dispatch(actions.fetchConfig())
-      .then(() => { // return of async actions
+      .then(() => {
         expect(store.getActions()).toEqual(expectedActions);
       });
   });
 
   it('creates PUT_CONFIG_FAILURE when updating configuration failed', () => {
-
-    let config = { title: 'Awesome Title' };
     nock(API)
       .put('/configuration', config)
       .replyWithError('something awful happened');
@@ -46,9 +46,8 @@ describe('Actions::Config', () => {
     const store = mockStore({ config: {} });
 
     return store.dispatch(actions.putConfig(config))
-      .then(() => { // return of async actions
+      .then(() => {
         expect(store.getActions()[0].type).toEqual(expectedAction.type);
       });
   });
-
 });
