+++ lib/jekyll-admin/public/src/actions/tests/datafiles.spec.js
@@ -174,5 +174,4 @@ describe('Actions::Datafiles', () => {
     store.dispatch(actions.putDataFile());
     expect(store.getActions()).toEqual(expectedActions);
   });
-
 });
