+++ lib/jekyll-admin/public/src/actions/tests/fixtures/index.js
@@ -1,3 +1,8 @@
+export const config = {
+  title: 'Awesome Title',
+  gems: 'jekyll-admin'
+};
+
 export const doc = {
   id: "/movies/the-revenant",
   ext: ".md",
