+++ lib/jekyll-admin/public/src/actions/tests/metadata.spec.js
@@ -13,6 +13,7 @@ describe('Actions::Metadata', () => {
     };
     expect(actions.storeContentFields(content)).toEqual(expectedAction);
   });
+
   it('creates ADD_METAFIELD', () => {
     const expectedAction = {
       type: types.ADD_METAFIELD,
@@ -20,6 +21,7 @@ describe('Actions::Metadata', () => {
     };
     expect(actions.addField('metadata["mentors"]')).toEqual(expectedAction);
   });
+
   it('creates REMOVE_METAFIELD', () => {
     const expectedAction = {
       type: types.REMOVE_METAFIELD,
@@ -28,6 +30,7 @@ describe('Actions::Metadata', () => {
     };
     expect(actions.removeField('metadata', 'layout')).toEqual(expectedAction);
   });
+
   it('creates UPDATE_FIELD_KEY', () => {
     const expectedAction = {
       type: types.UPDATE_FIELD_KEY,
@@ -39,6 +42,7 @@ describe('Actions::Metadata', () => {
       actions.updateFieldKey('metadata', 'layout', 'layout1')
     ).toEqual(expectedAction);
   });
+
   it('creates UPDATE_FIELD_VALUE', () => {
     const expectedAction = {
       type: types.UPDATE_FIELD_VALUE,
@@ -49,6 +53,7 @@ describe('Actions::Metadata', () => {
       actions.updateFieldValue('metadata["layout"]', 'post1')
     ).toEqual(expectedAction);
   });
+
   it('creates MOVE_ARRAY_ITEM', () => {
     const expectedAction = {
       type: types.MOVE_ARRAY_ITEM,
@@ -60,6 +65,7 @@ describe('Actions::Metadata', () => {
       actions.moveArrayItem('metadata["mentors"]', 0, 1)
     ).toEqual(expectedAction);
   });
+
   it('creates CONVERT_FIELD', () => {
     const expectedAction = {
       type: types.CONVERT_FIELD,
@@ -70,33 +76,31 @@ describe('Actions::Metadata', () => {
       actions.convertField('metadata["mentors"]', 'simple')
     ).toEqual(expectedAction);
   });
+
   it('creates UPDATE_TITLE', () => {
     const expectedAction = {
       type: types.UPDATE_TITLE,
       title: 'Test Title'
     };
-    expect(
-      actions.updateTitle('Test Title')
-    ).toEqual(expectedAction);
+    expect(actions.updateTitle('Test Title')).toEqual(expectedAction);
   });
+
   it('creates UPDATE_BODY', () => {
     const expectedAction = {
       type: types.UPDATE_BODY,
       body: 'Test Body'
     };
-    expect(
-      actions.updateBody('Test Body')
-    ).toEqual(expectedAction);
+    expect(actions.updateBody('Test Body')).toEqual(expectedAction);
   });
+
   it('creates UPDATE_DRAFT', () => {
     const expectedAction = {
       type: types.UPDATE_DRAFT,
       draft: false
     };
-    expect(
-      actions.updateDraft(false)
-    ).toEqual(expectedAction);
+    expect(actions.updateDraft(false)).toEqual(expectedAction);
   });
+
   it('creates UPDATE_PATH', () => {
     const expectedAction = {
       type: types.UPDATE_PATH,
