+++ lib/jekyll-admin/public/src/actions/tests/pages.spec.js
@@ -16,7 +16,7 @@ describe('Actions::Pages', () => {
     nock.cleanAll();
   });
 
-  it('creates FETCH_PAGES_SUCCESS when fetching pages has been done', () => {
+  it('fetches pages successfully', () => {
     nock(API)
       .get('/pages')
       .reply(200, [page]);
@@ -34,10 +34,9 @@ describe('Actions::Pages', () => {
       });
   });
 
-  it('fetches the page', () => {
-    let id = 'about';
+  it('fetches the page successfully', () => {
     nock(API)
-      .get(`/pages/${id}`)
+      .get(`/pages/${page.name}`)
       .reply(200, page);
 
     const expectedActions = [
@@ -47,13 +46,13 @@ describe('Actions::Pages', () => {
 
     const store = mockStore({ page: {}, isFetching: true });
 
-    return store.dispatch(actions.fetchPage(id))
+    return store.dispatch(actions.fetchPage(page.name))
       .then(() => {
         expect(store.getActions()).toEqual(expectedActions);
       });
   });
 
-  it('deletes the page', () => {
+  it('deletes the page successfully', () => {
     nock(API)
       .delete(`/pages/${page.name}`)
       .reply(200, { message: 'Deleted' });
@@ -70,7 +69,7 @@ describe('Actions::Pages', () => {
       });
   });
 
-  it('creates DELETE_PAGE_FAILURE when deleting page failed', () => {
+  it('creates DELETE_PAGE_FAILURE when deleting a page failed', () => {
     nock(API)
       .delete(`/pages/${page.name}`)
       .replyWithError('something awful happened');
@@ -88,7 +87,7 @@ describe('Actions::Pages', () => {
       });
   });
 
-  it('updates the page', () => {
+  it('updates the page successfully', () => {
     nock(API)
       .put(`/pages/${page.name}`)
       .reply(200, page);
@@ -106,7 +105,7 @@ describe('Actions::Pages', () => {
       });
   });
 
-  it('creates the page', () => {
+  it('creates the page successfully', () => {
     nock(API)
       .put(`/pages/${new_page.path}`)
       .reply(200, page);
@@ -157,5 +156,4 @@ describe('Actions::Pages', () => {
     store.dispatch(actions.putPage(page.name));
     expect(store.getActions()).toEqual(expectedActions);
   });
-
 });
