+++ lib/jekyll-admin/public/src/actions/tests/staticfiles.spec.js
@@ -118,5 +118,4 @@ describe('Actions::StaticFiles', () => {
         expect(store.getActions()[0].type).toEqual(expectedActions[0].type);
       });
   });
-
 });
