+++ lib/jekyll-admin/public/src/components/Breadcrumbs.js
@@ -1,5 +1,4 @@
 import React, { Component, PropTypes } from 'react';
-import { findDOMNode } from 'react-dom';
 import { Link } from 'react-router';
 import _ from 'underscore';
