+++ lib/jekyll-admin/public/src/components/FilePreview.js
@@ -4,7 +4,7 @@ export default class FilePreview extends Component {
 
   handleClickDelete(path) {
     const { onClickDelete } = this.props;
-    let confirm = window.confirm("Are you sure that you want to delete this file?");
+    const confirm = window.confirm("Are you sure that you want to delete this file?");
     if (confirm) {
       const filename = path.substring(path.lastIndexOf('/') + 1);
       onClickDelete(filename);
