+++ lib/jekyll-admin/public/src/constants/actionTypes.js
@@ -87,7 +87,6 @@ export const DELETE_DATAFILE_FAILURE = 'DELETE_DATAFILE_FAILURE';
 export const DATAFILE_CHANGED = 'DATAFILE_CHANGED';
 
 // Static Files
-
 export const FETCH_STATICFILES_REQUEST = 'FETCH_STATICFILES_REQUEST';
 export const FETCH_STATICFILES_SUCCESS = 'FETCH_STATICFILES_SUCCESS';
 export const FETCH_STATICFILES_FAILURE = 'FETCH_STATICFILES_FAILURE';
