+++ lib/jekyll-admin/public/src/containers/Header.js
@@ -16,12 +16,11 @@ export class Header extends Component {
 
   render() {
     const { config } = this.props;
-    let title = _.isEmpty(config) ? '' : config.title;
     return (
       <div className="header">
         <h3 className="title">
           <Link target="_blank" to={`/`}>
-            <i className="fa fa-home"></i> <span>{title}</span>
+            <i className="fa fa-home"></i> <span>{config.title || 'You have no title!'}</span>
           </Link>
         </h3>
       </div>
