+++ lib/jekyll-admin/public/src/containers/MetaFields.js
@@ -29,7 +29,7 @@ export class MetaFields extends Component {
 
     const { path, title, raw_content, ...rest } = metadata;
 
-    let metafields = _.map(rest, (field, key) => {
+    const metafields = _.map(rest, (field, key) => {
       let type = "simple";
       if (_.isObject(field)) type = "object";
       if (_.isArray(field)) type = "array";
@@ -61,7 +61,6 @@ export class MetaFields extends Component {
       </div>
     );
   }
-
 }
 
 MetaFields.propTypes = {
