+++ lib/jekyll-admin/public/src/containers/Sidebar.js
@@ -24,8 +24,11 @@ export class Sidebar extends Component {
 
   renderCollections() {
     const { collections } = this.props;
-    if (collections.length) {
-      return _.chain(collections)
+    if (!collections.length) {
+      return null;
+    }
+
+    return _.chain(collections)
       .filter(col => col.label != 'posts') // TODO
       .map((col, i) => {
         return (
@@ -36,7 +39,6 @@ export class Sidebar extends Component {
           </li>
         );
       }).value();
-    }
   }
 
   render() {
