+++ lib/jekyll-admin/public/src/containers/tests/fixtures/index.js
@@ -1,12 +1,9 @@
 export const config = {
-  "title": "Your awesome title",
-  "email": "your-email@domain.com",
-  "description": "Write an awesome description for your new site here. You can edit this line in _config.yml. It will appear in your document head meta (for Google search results) and in your feed.xml site description.\n",
-  "baseurl": "",
-  "url": "http://yourdomain.com",
-  "twitter_username": "jekyllrb",
-  "github_username": "jekyll",
-  "markdown": "kramdown"
+  title: "Your awesome title",
+  email: "your-email@domain.com",
+  description: "Write an awesome description for your new site here.",
+  baseurl: "",
+  url: "http://yourdomain.com"
 };
 
 export const collections = [
