+++ lib/jekyll-admin/public/src/containers/tests/header.spec.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import expect from 'expect';
-import { Link } from 'react-router';
 import { Header } from '../Header';
 
 import { config } from './fixtures';
@@ -11,7 +10,7 @@ function setup() {
     fetchConfig: expect.createSpy()
   };
 
-  let component = mount(
+  const component = mount(
     <Header config={config} {...actions} />
   );
 
@@ -34,5 +33,4 @@ describe('Containers::Header', () => {
     const { actions } = setup();
     expect(actions.fetchConfig).toHaveBeenCalled();
   });
-
 });
