+++ lib/jekyll-admin/public/src/containers/tests/metafields.spec.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import { Link } from 'react-router';
 import expect from 'expect';
 import { mount } from 'enzyme';
 import _ from 'underscore';
@@ -20,7 +19,7 @@ function setup() {
     convertField: expect.createSpy()
   };
 
-  let component = mount(
+  const component = mount(
     <MetaFields
       fields={content}
       metadata={content}
