+++ lib/jekyll-admin/public/src/containers/tests/sidebar.spec.js
@@ -11,7 +11,7 @@ function setup() {
     fetchCollections: expect.createSpy()
   };
 
-  let component = mount(
+  const component = mount(
     <Sidebar collections={collections} {...actions} />
   );
 
@@ -35,5 +35,4 @@ describe('Containers::Sidebar', () => {
     const { actions } = setup();
     expect(actions.fetchCollections).toHaveBeenCalled();
   });
-
 });
