+++ lib/jekyll-admin/public/src/containers/views/Configuration.js
@@ -7,15 +7,16 @@ import Editor from '../../components/Editor';
 
 // Actions
 import { putConfig, onEditorChange } from '../../actions/config';
-import { toYAML, toJSON } from '../../utils/helpers';
+
+import { toJSON } from '../../utils/helpers';
 
 export class Configuration extends Component {
 
   handleSaveClick() {
     const { editorChanged, putConfig } = this.props;
     if (editorChanged) {
-      let value = this.refs.editor.getValue();
-      let json = toJSON(value);
+      const value = this.refs.editor.getValue();
+      const json = toJSON(value);
       putConfig(json);
     }
   }
