+++ lib/jekyll-admin/public/src/containers/views/DataFileEdit.js
@@ -31,14 +31,14 @@ export class DataFileEdit extends Component {
   handleClickSave() {
     const { datafileChanged, putDataFile, params } = this.props;
     if (datafileChanged) {
-      let value = this.refs.editor.getValue();
+      const value = this.refs.editor.getValue();
       putDataFile(params.data_file, value);
     }
   }
 
   handleClickDelete(filename) {
     const { deleteDataFile } = this.props;
-    let confirm = window.confirm("Are you sure that you want to delete this file?");
+    const confirm = window.confirm(`Are you sure that you want to delete "${filename}" ?`);
     if (confirm) {
       deleteDataFile(filename);
       browserHistory.push(`${ADMIN_PREFIX}/datafiles`);
