+++ lib/jekyll-admin/public/src/containers/views/DataFileNew.js
@@ -24,7 +24,7 @@ export class DataFileNew extends Component {
 
   componentWillReceiveProps(nextProps) {
     if (this.props.updated !== nextProps.updated) {
-      let filename = this.refs.breadcrumbs.refs.input.value;
+      const filename = this.refs.breadcrumbs.refs.input.value;
       browserHistory.push(`${ADMIN_PREFIX}/datafiles/${filename}`);
     }
   }
