+++ lib/jekyll-admin/public/src/containers/views/DataFiles.js
@@ -26,7 +26,7 @@ export class DataFiles extends Component {
 
   handleClickDelete(filename) {
     const { deleteDataFile } = this.props;
-    let confirm = window.confirm("Are you sure that you want to delete this file?");
+    const confirm = window.confirm(`Are you sure that you want to delete "${filename}" ?`);
     if (confirm) {
       deleteDataFile(filename);
     }
