+++ lib/jekyll-admin/public/src/containers/views/DocumentEdit.js
@@ -44,7 +44,7 @@ export class DocumentEdit extends Component {
 
   handleClickDelete(filename, collection) {
     const { deleteDocument } = this.props;
-    let confirm = window.confirm("Are you sure that you want to delete this document?");
+    const confirm = window.confirm(`Are you sure that you want to delete "${filename}" ?`);
     if (confirm) {
       deleteDocument(filename, collection);
       browserHistory.push(`${ADMIN_PREFIX}/collections/${collection}`);
