+++ lib/jekyll-admin/public/src/containers/views/Documents.js
@@ -35,7 +35,7 @@ export class Documents extends Component {
 
   handleClickDelete(filename, collection) {
     const { deleteDocument } = this.props;
-    let confirm = window.confirm("Are you sure that you want to delete this document?");
+    const confirm = window.confirm(`Are you sure that you want to delete "${filename}" ?`);
     if (confirm) {
       deleteDocument(filename, collection);
     }
