+++ lib/jekyll-admin/public/src/containers/views/NotFound.js
@@ -8,7 +8,6 @@ export default class NotFound extends Component {
         <h1>
           Huh. It seems that page is Hyde-ing...
         </h1>
-
         <h2>The resource you requested was not found.</h2>
       </div>
     );
