+++ lib/jekyll-admin/public/src/containers/views/PageEdit.js
@@ -41,7 +41,7 @@ export class PageEdit extends Component {
 
   handleClickDelete(name) {
     const { deletePage } = this.props;
-    let confirm = window.confirm("Are you sure that you want to delete this page?");
+    const confirm = window.confirm(`Are you sure that you want to delete "${name}"?`);
     if (confirm) {
       deletePage(name);
       browserHistory.push(`${ADMIN_PREFIX}/pages`);
