+++ lib/jekyll-admin/public/src/containers/views/Pages.js
@@ -26,7 +26,7 @@ export class Pages extends Component {
 
   handleClickDelete(name) {
     const { deletePage } = this.props;
-    let confirm = window.confirm("Are you sure that you want to delete this page?");
+    const confirm = window.confirm(`Are you sure that you want to delete "${name}"?`);
     if (confirm) {
       deletePage(name);
     }
