+++ lib/jekyll-admin/public/src/containers/views/StaticFiles.js
@@ -2,13 +2,13 @@ import React, { Component, PropTypes } from 'react';
 import { Link } from 'react-router';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import Dropzone from 'react-dropzone';
 import _ from 'underscore';
 
 // Constants
 import { ADMIN_PREFIX } from '../../constants';
 
 // Components
+import Dropzone from 'react-dropzone';
 import FilePreview from '../../components/FilePreview';
 import InputSearch from '../../components/form/InputSearch';
