+++ lib/jekyll-admin/public/src/containers/views/tests/configuration.spec.js
@@ -1,8 +1,9 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import expect from 'expect';
-import { Configuration } from '../Configuration';
 import Editor from '../../../components/Editor';
+
+import { Configuration } from '../Configuration';
 
 import { config } from './fixtures';
 
@@ -16,7 +17,7 @@ const defaultProps = {
 };
 
 function setup(props = defaultProps) {
-  let component = shallow(<Configuration {...props} />);
+  const component = shallow(<Configuration {...props} />);
 
   return {
     component,
@@ -35,6 +36,7 @@ describe('Containers::Configuration', () => {
     expect(message.node).toNotExist();
     expect(editor.prop('json')).toEqual(config);
   });
+
   it('should render correctly with updated props', () => {
     const { component, editor, saveButton, message } = setup(
       Object.assign({}, defaultProps, {
