+++ lib/jekyll-admin/public/src/containers/views/tests/datafileedit.spec.js
@@ -24,7 +24,7 @@ function setup(props = defaultProps) {
     clearErrors: expect.createSpy()
   };
 
-  let component = shallow(
+  const component = shallow(
     <DataFileEdit {...actions} {...props} />
   );
