+++ lib/jekyll-admin/public/src/containers/views/tests/datafilenew.spec.js
@@ -21,7 +21,7 @@ function setup(props = defaultProps) {
     clearErrors: expect.createSpy()
   };
 
-  let component = shallow(
+  const component = shallow(
     <DataFileNew {...actions} {...props} />
   );
 
@@ -39,6 +39,7 @@ describe('Containers::DataFileNew', () => {
     saveButton.simulate('click');
     expect(actions.putDataFile).toNotHaveBeenCalled();
   });
+
   it('should call activate save button if a field is changed.', () => {
     const { saveButton, actions } = setup(Object.assign({}, defaultProps, {
       datafileChanged: true
