+++ lib/jekyll-admin/public/src/containers/views/tests/datafiles.spec.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import expect from 'expect';
+
 import { DataFiles } from '../DataFiles';
 
 import { datafile } from './fixtures';
@@ -12,7 +13,7 @@ function setup(datafiles=[datafile]) {
     search: expect.createSpy()
   };
 
-  let component = mount(
+  const component = mount(
     <DataFiles
       files={datafiles}
       message=""
