+++ lib/jekyll-admin/public/src/containers/views/tests/documentedit.spec.js
@@ -28,7 +28,7 @@ function setup(props = defaultProps) {
     clearErrors: expect.createSpy()
   };
 
-  let component = shallow(
+  const component = shallow(
     <DocumentEdit {...actions} {...props} />
   );
 
@@ -47,6 +47,7 @@ describe('Containers::DocumentEdit', () => {
     const { actions } = setup();
     expect(actions.clearErrors).toHaveBeenCalled();
   });
+
   it('should render correctly', () => {
     let { component } = setup(Object.assign(
       {}, defaultProps, { isFetching: true }
@@ -56,21 +57,25 @@ describe('Containers::DocumentEdit', () => {
     )).component;
     expect(component.find('h1').node).toExist();
   });
+
   it('should not render error messages initially', () => {
     const { errors } = setup();
     expect(errors.node).toNotExist();
   });
+
   it('should render error messages', () => {
     const { errors } = setup(Object.assign({}, defaultProps, {
       errors: ['The title field is required!']
     }));
     expect(errors.node).toExist();
   });
+
   it('should not call putDocument if a field is not changed.', () => {
     const { saveButton, actions } = setup();
     saveButton.simulate('click');
     expect(actions.putDocument).toNotHaveBeenCalled();
   });
+
   it('should call putDocument if a field is changed.', () => {
     const { saveButton, actions } = setup(Object.assign({}, defaultProps, {
       fieldChanged: true
@@ -78,6 +83,7 @@ describe('Containers::DocumentEdit', () => {
     saveButton.simulate('click');
     expect(actions.putDocument).toHaveBeenCalled();
   });
+
   it('should call deleteDocument', () => {
     const { deleteButton, actions } = setup();
     deleteButton.simulate('click');
