+++ lib/jekyll-admin/public/src/containers/views/tests/documentnew.spec.js
@@ -24,7 +24,7 @@ function setup(props = defaultProps) {
     clearErrors: expect.createSpy()
   };
 
-  let component = shallow(
+  const component = shallow(
     <DocumentNew {...actions} {...props} />
   );
 
@@ -40,24 +40,27 @@ function setup(props = defaultProps) {
 describe('Containers::DocumentNew', () => {
   it('should call clearErrors before mount', () => {
     const { actions } = setup();
-
     expect(actions.clearErrors).toHaveBeenCalled();
   });
+
   it('should not render error messages initially', () => {
     const { errors } = setup();
     expect(errors.node).toNotExist();
   });
+
   it('should render error messages', () => {
     const { errors } = setup(Object.assign({}, defaultProps, {
       errors: ['The title field is required!']
     }));
     expect(errors.node).toExist();
   });
+
   it('should not call putDocument if a field is not changed.', () => {
     const { saveButton, actions } = setup();
     saveButton.simulate('click');
     expect(actions.putDocument).toNotHaveBeenCalled();
   });
+
   it('should call putDocument if a field is changed.', () => {
     const { saveButton, actions } = setup(Object.assign({}, defaultProps, {
       fieldChanged: true
