+++ lib/jekyll-admin/public/src/containers/views/tests/documents.spec.js
@@ -13,7 +13,7 @@ function setup(currentDocuments=[doc]) {
     search: expect.createSpy()
   };
 
-  let component = mount(
+  const component = mount(
     <Documents
       currentDocuments={currentDocuments}
       message=""
