+++ lib/jekyll-admin/public/src/containers/views/tests/pagenew.spec.js
@@ -23,7 +23,7 @@ function setup(props = defaultProps) {
     clearErrors: expect.createSpy()
   };
 
-  let component = shallow(
+  const component = shallow(
     <PageNew {...actions} {...props} />
   );
 
@@ -41,21 +41,25 @@ describe('Containers::PageNew', () => {
     const { actions } = setup();
     expect(actions.clearErrors).toHaveBeenCalled();
   });
+
   it('should not render error messages initially', () => {
     const { errors } = setup();
     expect(errors.node).toNotExist();
   });
+
   it('should render error messages', () => {
     const { errors } = setup(Object.assign({}, defaultProps, {
       errors: ['The path field is required!']
     }));
     expect(errors.node).toExist();
   });
+
   it('should not call putPage if a field is not changed.', () => {
     const { saveButton, actions } = setup();
     saveButton.simulate('click');
     expect(actions.putPage).toNotHaveBeenCalled();
   });
+
   it('should call putPage if a field is changed.', () => {
     const { saveButton, actions } = setup(Object.assign({}, defaultProps, {
       fieldChanged: true
