+++ lib/jekyll-admin/public/src/containers/views/tests/pages.spec.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import expect from 'expect';
+
 import { Pages } from '../Pages';
 
 import { page } from './fixtures';
@@ -12,7 +13,7 @@ function setup(pages=[page]) {
     search: expect.createSpy()
   };
 
-  let component = mount(
+  const component = mount(
     <Pages
       pages={pages}
       message=""
