+++ lib/jekyll-admin/public/src/containers/views/tests/staticfiles.spec.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import expect from 'expect';
+
 import { StaticFiles } from '../StaticFiles';
 
 import { staticfile } from './fixtures';
@@ -13,7 +14,7 @@ function setup(files=[staticfile]) {
     search: expect.createSpy()
   };
 
-  let component = mount(
+  const component = mount(
     <StaticFiles
       files={files}
       message=""
