+++ lib/jekyll-admin/public/src/reducers/metadata.js
@@ -8,7 +8,7 @@ import {
 } from '../constants/actionTypes';
 
 import { addField, removeField, updateFieldKey,
-  updateFieldValue, moveArrayItem, convertField } from '../utils/helpers';
+  updateFieldValue, moveArrayItem, convertField } from '../utils/metadata';
 
 // TODO normalize the metadata
 export default function metadata(state = {
@@ -50,11 +50,10 @@ export default function metadata(state = {
         },
         fieldChanged: true
       });
-    case STORE_CONTENT_FIELDS: {
+    case STORE_CONTENT_FIELDS:
       return Object.assign({}, state, {
         metadata: action.content
       });
-    }
     case ADD_METAFIELD:
       return Object.assign({}, state, {
         metadata: addField(state, action.namePrefix),
