+++ lib/jekyll-admin/public/src/reducers/tests/config.spec.js
@@ -89,5 +89,4 @@ describe('Reducers::Config', () => {
       updated: false
     });
   });
-
 });
