+++ lib/jekyll-admin/public/src/reducers/tests/datafiles.spec.js
@@ -136,5 +136,4 @@ describe('Reducers::DataFiles', () => {
       updated: false
     });
   });
-
 });
