+++ lib/jekyll-admin/public/src/reducers/tests/metadata.spec.js
@@ -241,5 +241,4 @@ describe('Reducers::Metadata', () => {
       fieldChanged: true
     });
   });
-
 });
