+++ lib/jekyll-admin/public/src/reducers/tests/utils.spec.js
@@ -48,5 +48,4 @@ describe('Reducers::Utils', () => {
       errors: ['The title is required']
     });
   });
-
 });
