+++ lib/jekyll-admin/public/src/routes.js
@@ -25,7 +25,7 @@ export default (
     <Route path="page/new" component={PageNew} />
     <Route path="collections/:collection_name" component={Documents} />
     <Route path="collections/:collection_name/:id" component={DocumentEdit} />
-    <Route path="collection/:collection_name/new" component={DocumentNew} /> // check collection exists
+    <Route path="collection/:collection_name/new" component={DocumentNew} /> // TODO check collection exists
     <Route path="datafiles" component={DataFiles} />
     <Route path="datafiles/:data_file" component={DataFileEdit} />
     <Route path="datafile/new" component={DataFileNew} />
