+++ lib/jekyll-admin/public/src/store/configureStore.dev.js
@@ -2,6 +2,7 @@ import { createStore, applyMiddleware, compose } from 'redux';
 import rootReducer from '../reducers';
 import thunk from 'redux-thunk';
 import createLogger from 'redux-logger';
+
 const logger = createLogger();
 
 export default function configureStore(initialState) {
