+++ lib/jekyll-admin/public/src/utils/tests/validation.spec.js
@@ -5,7 +5,7 @@ import { validator } from '../validation';
 describe('Validation functions:', () => {
   describe('validator', () => {
     it('should not generate error messages if all fields are valid', () => {
-      let metadata = {
+      const metadata = {
         title: 'Test Title',
         path: '2016-01-11-awesome-path-test.md'
       };
@@ -22,8 +22,9 @@ describe('Validation functions:', () => {
       );
       expect(errorMessages).toEqual([]);
     });
+
     it('should generate error messages if not validated', () => {
-      let metadata = {};
+      const metadata = {};
       const errorMessages = validator(
         metadata,
         {
@@ -40,8 +41,9 @@ describe('Validation functions:', () => {
         'The title is required.'
       ]);
     });
+
     it("should generate an error message if a post's filename does not include date", () => {
-      let metadata = {
+      const metadata = {
         path: 'test.md'
       };
       const errorMessages = validator(
